requirements.txt
# 基础框架
streamlit>=1.32.0
python-dotenv>=1.0.0

# LangChain 核心
langchain>=0.1.10
langchain-community>=0.0.28
langchain-core>=0.1.28
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1

# LangGraph（Agent核心）
langgraph>=0.0.41

# 阿里云通义千问
dashscope>=1.14.0

# 文档处理
pypdf>=4.0.0

# 向量库
faiss-cpu>=1.8.0

# 其他依赖
requests>=2.31.0
pydantic>=2.0.0
